refactor: split manager media router into ingest and gallery modules
